requirements.txt
# --- Core Backend Dependencies ---
fastapi
uvicorn
python-dotenv
openai
git+https://github.com/openai/whisper.git
spacy
presidio-analyzer
presidio-anonymizer
torchaudio

# --- Sentiment Analysis ---
transformers

# --- Analytics and Notebooks ---
pandas
duckdb
altair
jupyter

# --- ML and NLP ---
scikit-learn
xgboost
imbalanced-learn

# --- Miscellaneous ---
requests
pathlib
keras
streamlit
